ASNIDE-330 Extracted common "String" class
