+++ src/lib/data/types/ia5string.h
@@ -27,17 +27,14 @@
 
 #include <QString>
 
-#include "constraints.h"
-
+#include <data/types/string.h>
 #include <data/types/type.h>
 
 namespace MalTester {
 namespace Data {
 namespace Types {
 
-class IA5String : public Type,
-                  public WithConstraints<IntegerConstraints>,
-                  public WithConstraints<StringConstraints>
+class IA5String : public String
 {
 public:
     IA5String() = default;
@@ -46,26 +43,6 @@ public:
     QString name() const override { return QLatin1String("IA5String"); }
     void accept(TypeVisitor &visitor) override;
     std::unique_ptr<Type> clone() const override;
-
-    IntegerConstraints &integerConstraints()
-    {
-        return WithConstraints<IntegerConstraints>::constraints();
-    }
-
-    const IntegerConstraints &integerConstraints() const
-    {
-        return WithConstraints<IntegerConstraints>::constraints();
-    }
-
-    StringConstraints &stringConstraints()
-    {
-        return WithConstraints<StringConstraints>::constraints();
-    }
-
-    const StringConstraints &stringConstraints() const
-    {
-        return WithConstraints<StringConstraints>::constraints();
-    }
 };
 
 } // namespace Types
